feat : carousel 이벤트핸들러 연결
