+++ src/components/InfoWindow/Contents/Carousel/eventHandler/carouselChange.ts
@@ -1,0 +1,1 @@
+export const handleCarouselChange = (e: Event) => console.log(e);
