+++ src/components/InfoWindow/Contents/PopUpWindow/utils/cancelAnimation.ts
