+++ src/constants/DOMId.ts
@@ -1,0 +1,3 @@
+export const DOMID_POP_UP_WINDOW = 'pop_up_window';
+export const DOMID_BLURFRAME = 'blur_fram';
+export const DOMID_CAROUSEL = 'carousel';
