+++ src/constants/event.ts
@@ -1,0 +1,1 @@
+export const EVENT_CHANGE_CAROUSEL = 'carousel_change';
